opendal_service_redb

Struct Redb

pub struct Redb { /* private fields */ }

Redb service support.

Capabilities

Depending on its configuration and the backing system, this service can expose:

create_dir

stat

read

write

delete

list

copy

rename

presign

Inspect the effective capability set with [opendal_core::Operator::info] and [opendal_core::OperatorInfo::capability] after building an operator.

Configuration

Use crate::RedbConfig for serializable configuration and this builder’s methods for direct construction. The field and method documentation defines accepted values, defaults, and environment interaction.

Example

Via Builder

use opendal_core::Operator;
use opendal_core::Result;
use opendal_service_redb::Redb;

#[tokio::main]
async fn main() -> Result<()> {
    let mut builder = Redb::default()
        .datadir("/tmp/opendal/redb")
        .table("opendal-redb");

    let op: Operator = Operator::new(builder)?;
    Ok(())
}

Implementations

impl RedbBuilder

pub fn database(self, db: Arc<Database>) -> Self

Set the database for Redb.

This method should be called when you want to use multiple tables of one database because Redb doesn’t allow opening a database that have been opened.

datadir and database should not be set simultaneously. If both are set, database will take precedence.

pub fn datadir(self, path: &str) -> Self

Set the path to the redb data directory. Will create if not exists.

Opening redb database via datadir takes away the ability to access multiple redb tables. If you need to access multiple redb tables, the correct solution is to create an Arc<redb::database> beforehand and then share it via database with multiple builders where every builder will open one redb table.

pub fn table(self, table: &str) -> Self

Set the table name for Redb. Will create if not exists.

pub fn root(self, path: &str) -> Self

Set the root for Redb.